% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tether.R
\name{as_tether}
\alias{as_tether}
\alias{as_tether.default}
\alias{as_tether.connection}
\alias{as_tether.python.builtin.object}
\alias{as_tether.python.builtin.type}
\alias{as_tether.python.builtin.module}
\alias{as_tether.python.builtin.function}
\title{Generate tether}
\usage{
as_tether(x, ..., roxify = TRUE)

\method{as_tether}{default}(x, ..., roxify = TRUE)

\method{as_tether}{connection}(x, ..., roxify = TRUE)

\method{as_tether}{python.builtin.object}(x, ..., roxify = TRUE)

\method{as_tether}{python.builtin.type}(x, ..., roxify = TRUE, inherited = FALSE)

\method{as_tether}{python.builtin.module}(x, ..., roxify = TRUE)

\method{as_tether}{python.builtin.function}(x, ..., name = x$`__name__`, roxify = TRUE)
}
\arguments{
\item{x}{an object}

\item{...}{passed on to methods}

\item{roxify}{whether a attribute \code{"roxified"} should be included. This is
the value that will be used to rebase the roxygen block instead of the
tether itself.}

\item{inherited}{Whether to include inherited methods and attributes.}

\item{name}{the name to present in the Python \verb{__signature__}}
}
\value{
the tether, a string or character vector (of lines) that will be used
to track changes and rebase the documentation. If \code{roxify} is \code{TRUE}, the
returned tether will contain an attribute \code{attr(tether, "roxified")}, which
will also be a string. If present, \code{"roxified"} will be used to perform the
rebase.
}
\description{
Generate tether
}
