% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retether.R
\name{roxy_tag_parse.roxy_tag_tether}
\alias{roxy_tag_parse.roxy_tag_tether}
\title{Stub tag parse for \verb{@tether} tags}
\usage{
\method{roxy_tag_parse}{roxy_tag_tether}(x)
}
\arguments{
\item{x}{The \verb{@tether} tag}
}
\value{
\code{x}, unmodified
}
\description{
This method is exported to allow roxygen2 to ignore the \verb{@tether} tag
during the standard \code{roxygen2::roxygenise()} workflow.
}
\keyword{internal}
