% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retether.R, R/roclet.R
\name{retether}
\alias{retether}
\title{Retether an R package}
\usage{
retether(base_path = ".", tag_parser = NULL, resolve_tether_file = NULL)

retether(base_path = ".", tag_parser = NULL, resolve_tether_file = NULL)
}
\arguments{
\item{base_path}{path to an R package.}

\item{tag_parser}{A function taking a single argument, the raw \verb{@tether}
string, returning a tether. If \code{NULL}, (the default), the tag value is
evaluated with \code{eval(str2expression(tag$raw), globalenv())}}

\item{resolve_tether_file}{if \code{NULL}, the default, tether files are stored
under \code{"man-src/tether/{block_name}.R"}. If supplied a string, this is
taken as the directory where to store tether files (instead of
\code{"man-src/tether"}). If a function, it is expected to take a single
argument, the block name, and return a filepath.}
}
\description{
Retether an R package

Retether an R package
}
